Revert "Flutter Web Driver Support (#45951)" (#47389)
This reverts commit 1ef0eadb45835dc438fe0120521c01c8b055b41e.
