+++ packages/flutter_driver/lib/src/driver/driver.dart
@@ -3,15 +3,23 @@
 // found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:convert';
+import 'dart:io';
 
+import 'package:file/file.dart' as f;
+import 'package:fuchsia_remote_debug_protocol/fuchsia_remote_debug_protocol.dart' as fuchsia;
+import 'package:json_rpc_2/error_code.dart' as error_code;
 import 'package:json_rpc_2/json_rpc_2.dart' as rpc;
 import 'package:meta/meta.dart';
+import 'package:path/path.dart' as p;
 import 'package:vm_service_client/vm_service_client.dart';
+import 'package:web_socket_channel/io.dart';
 
 import '../common/diagnostics_tree.dart';
 import '../common/error.dart';
 import '../common/find.dart';
 import '../common/frame_sync.dart';
+import '../common/fuchsia_compat.dart';
 import '../common/geometry.dart';
 import '../common/gesture.dart';
 import '../common/health.dart';
@@ -22,12 +30,8 @@ import '../common/request_data.dart';
 import '../common/semantics.dart';
 import '../common/text.dart';
 import '../common/wait.dart';
+import 'common.dart';
 import 'timeline.dart';
-import 'vmservice_driver.dart';
-import 'web_driver.dart';
-
-export 'vmservice_driver.dart';
-export 'web_driver.dart';
 
 /// Timeline stream identifier.
 enum TimelineStream {
@@ -59,10 +63,51 @@ enum TimelineStream {
   vm,
 }
 
+const List<TimelineStream> _defaultStreams = <TimelineStream>[TimelineStream.all];
+
 /// How long to wait before showing a message saying that
 /// things seem to be taking a long time.
 @visibleForTesting
 const Duration kUnusuallyLongTimeout = Duration(seconds: 5);
+
+/// The amount of time we wait prior to making the next attempt to connect to
+/// the VM service.
+const Duration _kPauseBetweenReconnectAttempts = Duration(seconds: 1);
+
+// See https://github.com/dart-lang/sdk/blob/master/runtime/vm/timeline.cc#L32
+String _timelineStreamsToString(List<TimelineStream> streams) {
+  final String contents = streams.map<String>((TimelineStream stream) {
+    switch (stream) {
+      case TimelineStream.all: return 'all';
+      case TimelineStream.api: return 'API';
+      case TimelineStream.compiler: return 'Compiler';
+      case TimelineStream.dart: return 'Dart';
+      case TimelineStream.debugger: return 'Debugger';
+      case TimelineStream.embedder: return 'Embedder';
+      case TimelineStream.gc: return 'GC';
+      case TimelineStream.isolate: return 'Isolate';
+      case TimelineStream.vm: return 'VM';
+      default:
+        throw 'Unknown timeline stream $stream';
+    }
+  }).join(', ');
+  return '[$contents]';
+}
+
+void _log(String message) {
+  driverLog('FlutterDriver', message);
+}
+
+Future<T> _warnIfSlow<T>({
+  @required Future<T> future,
+  @required Duration timeout,
+  @required String message,
+}) {
+  assert(future != null);
+  assert(timeout != null);
+  assert(message != null);
+  return future..timeout(timeout, onTimeout: () { _log(message); return null; });
+}
 
 /// A convenient accessor to frequently used finders.
 ///
@@ -80,27 +125,43 @@ const CommonFinders find = CommonFinders._();
 typedef EvaluatorFunction = dynamic Function();
 
 /// Drives a Flutter Application running in another process.
-abstract class FlutterDriver {
-  /// Default constructor.
+class FlutterDriver {
+  /// Creates a driver that uses a connection provided by the given
+  /// [serviceClient], [_peer] and [appIsolate].
   @visibleForTesting
-  FlutterDriver();
+  FlutterDriver.connectedTo(
+    this.serviceClient,
+    this._peer,
+    this.appIsolate, {
+    bool printCommunication = false,
+    bool logCommunicationToFile = true,
+  }) : _printCommunication = printCommunication,
+       _logCommunicationToFile = logCommunicationToFile,
+       _driverId = _nextDriverId++;
 
-  /// Creates a driver that uses a connection provided by either the combination
-  /// of [webConnection] and [browser], or the combination of [serviceClient],
-  /// [peer] and [appIsolate]
-  @visibleForTesting
-  factory FlutterDriver.connectedTo({
-    FlutterWebConnection webConnection,
-    Browser browser,
-    VMServiceClient serviceClient,
-    rpc.Peer peer,
-    VMIsolate appIsolate,
-  }) {
-    if (webConnection != null && browser != null) {
-      return WebFlutterDriver.connectedTo(webConnection, browser);
-    }
-    return VMServiceFlutterDriver.connectedTo(serviceClient, peer, appIsolate);
-  }
+  static const String _flutterExtensionMethodName = 'ext.flutter.driver';
+  static const String _setVMTimelineFlagsMethodName = 'setVMTimelineFlags';
+  static const String _getVMTimelineMethodName = 'getVMTimeline';
+  static const String _clearVMTimelineMethodName = 'clearVMTimeline';
+  static const String _collectAllGarbageMethodName = '_collectAllGarbage';
+
+  static int _nextDriverId = 0;
+
+  // The additional blank line in the beginning is for _log.warning.
+  static const String _kDebugWarning = '''
+
+┏╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍┓
+┇ ⚠    THIS BENCHMARK IS BEING RUN IN DEBUG MODE     ⚠  ┇
+┡╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍╍┦
+│                                                       │
+│  Numbers obtained from a benchmark while asserts are  │
+│  enabled will not accurately reflect the performance  │
+│  that will be experienced by end users using release  ╎
+│  builds. Benchmarks should be run using this command  ┆
+│  line:  flutter drive --profile test_perf.dart        ┊
+│                                                       ┊
+└─────────────────────────────────────────────────╌┄┈  🐢
+''';
 
   /// Connects to a Flutter application.
   ///
@@ -127,10 +188,6 @@ abstract class FlutterDriver {
   /// `isolateNumber` is set, as this is already enough information to connect
   /// to an isolate.
   ///
-  /// `browser` specifies which FlutterDriver implementation to use. If not
-  /// speicifed or set to false, [VMServiceFlutterDriver] implementation
-  /// will be used. Otherwise, [WebFlutterDriver] implementation will be used.
-  ///
   /// The return value is a future. This method never times out, though it may
   /// fail (completing with an error). A timeout can be applied by the caller
   /// using [Future.timeout] if necessary.
@@ -140,64 +197,308 @@ abstract class FlutterDriver {
     bool logCommunicationToFile = true,
     int isolateNumber,
     Pattern fuchsiaModuleTarget,
-    bool browser = false,
-    Duration timeout,
   }) async {
-
-    if (browser) {
-      return WebFlutterDriver.connectWeb(hostUrl: dartVmServiceUrl, timeout: timeout);
+    // If running on a Fuchsia device, connect to the first isolate whose name
+    // matches FUCHSIA_MODULE_TARGET.
+    //
+    // If the user has already supplied an isolate number/URL to the Dart VM
+    // service, then this won't be run as it is unnecessary.
+    if (Platform.isFuchsia && isolateNumber == null) {
+      // On Fuchsia stderr/stdout appear to have issues working correctly,
+      // so we work around the issue by using print directly.
+      // TODO(awdavies): Fix Dart or Fuchsia to not need this workaround.
+      driverLog = (String source, String message) {
+        print('$source: $message');
+      };
+      fuchsiaModuleTarget ??= Platform.environment['FUCHSIA_MODULE_TARGET'];
+      if (fuchsiaModuleTarget == null) {
+        throw DriverError(
+          'No Fuchsia module target has been specified.\n'
+          'Please make sure to specify the FUCHSIA_MODULE_TARGET '
+          'environment variable.'
+        );
+      }
+      final fuchsia.FuchsiaRemoteConnection fuchsiaConnection =
+          await FuchsiaCompat.connect();
+      final List<fuchsia.IsolateRef> refs =
+          await fuchsiaConnection.getMainIsolatesByPattern(fuchsiaModuleTarget);
+      final fuchsia.IsolateRef ref = refs.first;
+      isolateNumber = ref.number;
+      dartVmServiceUrl = ref.dartVm.uri.toString();
+      await fuchsiaConnection.stop();
+      FuchsiaCompat.cleanup();
     }
-    return VMServiceFlutterDriver.connect(
-              dartVmServiceUrl: dartVmServiceUrl,
-              printCommunication: printCommunication,
-              logCommunicationToFile: logCommunicationToFile,
-              isolateNumber: isolateNumber,
-              fuchsiaModuleTarget: fuchsiaModuleTarget,
+
+    dartVmServiceUrl ??= Platform.environment['VM_SERVICE_URL'];
+
+    if (dartVmServiceUrl == null) {
+      throw DriverError(
+        'Could not determine URL to connect to application.\n'
+        'Either the VM_SERVICE_URL environment variable should be set, or an explicit '
+        'URL should be provided to the FlutterDriver.connect() method.'
+      );
+    }
+
+    // Connect to Dart VM services
+    _log('Connecting to Flutter application at $dartVmServiceUrl');
+    final VMServiceClientConnection connection =
+        await vmServiceConnectFunction(dartVmServiceUrl);
+    final VMServiceClient client = connection.client;
+    final VM vm = await client.getVM();
+    final VMIsolateRef isolateRef = isolateNumber ==
+        null ? vm.isolates.first :
+               vm.isolates.firstWhere(
+                   (VMIsolateRef isolate) => isolate.number == isolateNumber);
+    _log('Isolate found with number: ${isolateRef.number}');
+
+    VMIsolate isolate = await isolateRef.loadRunnable();
+
+    // TODO(yjbanov): vm_service_client does not support "None" pause event yet.
+    // It is currently reported as null, but we cannot rely on it because
+    // eventually the event will be reported as a non-null object. For now,
+    // list all the events we know about. Later we'll check for "None" event
+    // explicitly.
+    //
+    // See: https://github.com/dart-lang/vm_service_client/issues/4
+    if (isolate.pauseEvent is! VMPauseStartEvent &&
+        isolate.pauseEvent is! VMPauseExitEvent &&
+        isolate.pauseEvent is! VMPauseBreakpointEvent &&
+        isolate.pauseEvent is! VMPauseExceptionEvent &&
+        isolate.pauseEvent is! VMPauseInterruptedEvent &&
+        isolate.pauseEvent is! VMResumeEvent) {
+      isolate = await isolateRef.loadRunnable();
+    }
+
+    final FlutterDriver driver = FlutterDriver.connectedTo(
+      client, connection.peer, isolate,
+      printCommunication: printCommunication,
+      logCommunicationToFile: logCommunicationToFile,
     );
+
+    driver._dartVmReconnectUrl = dartVmServiceUrl;
+
+    // Attempts to resume the isolate, but does not crash if it fails because
+    // the isolate is already resumed. There could be a race with other tools,
+    // such as a debugger, any of which could have resumed the isolate.
+    Future<dynamic> resumeLeniently() {
+      _log('Attempting to resume isolate...');
+      return isolate.resume().catchError((dynamic e) {
+        const int vmMustBePausedCode = 101;
+        if (e is rpc.RpcException && e.code == vmMustBePausedCode) {
+          // No biggie; something else must have resumed the isolate
+          _log(
+            'Attempted to resume an already resumed isolate. This may happen '
+            'when we lose a race with another tool (usually a debugger) that '
+            'is connected to the same isolate.'
+          );
+        } else {
+          // Failed to resume due to another reason. Fail hard.
+          throw e;
+        }
+      });
+    }
+
+    /// Waits for a signal from the VM service that the extension is registered.
+    /// Returns [_flutterExtensionMethodName]
+    Future<String> waitForServiceExtension() {
+      return isolate.onExtensionAdded.firstWhere((String extension) {
+        return extension == _flutterExtensionMethodName;
+      });
+    }
+
+    /// Tells the Dart VM Service to notify us about "Isolate" events.
+    ///
+    /// This is a workaround for an issue in package:vm_service_client, which
+    /// subscribes to the "Isolate" stream lazily upon subscription, which
+    /// results in lost events.
+    ///
+    /// Details: https://github.com/dart-lang/vm_service_client/issues/17
+    Future<void> enableIsolateStreams() async {
+      await connection.peer.sendRequest('streamListen', <String, String>{
+        'streamId': 'Isolate',
+      });
+    }
+
+    // Attempt to resume isolate if it was paused
+    if (isolate.pauseEvent is VMPauseStartEvent) {
+      _log('Isolate is paused at start.');
+
+      // If the isolate is paused at the start, e.g. via the --start-paused
+      // option, then the VM service extension is not registered yet. Wait for
+      // it to be registered.
+      await enableIsolateStreams();
+      final Future<String> whenServiceExtensionReady = waitForServiceExtension();
+      final Future<dynamic> whenResumed = resumeLeniently();
+      await whenResumed;
+
+      _log('Waiting for service extension...');
+      // We will never receive the extension event if the user does not
+      // register it. If that happens, show a message but continue waiting.
+      await _warnIfSlow<String>(
+        future: whenServiceExtensionReady,
+        timeout: kUnusuallyLongTimeout,
+        message: 'Flutter Driver extension is taking a long time to become available. '
+                 'Ensure your test app (often "lib/main.dart") imports '
+                 '"package:flutter_driver/driver_extension.dart" and '
+                 'calls enableFlutterDriverExtension() as the first call in main().',
+      );
+    } else if (isolate.pauseEvent is VMPauseExitEvent ||
+               isolate.pauseEvent is VMPauseBreakpointEvent ||
+               isolate.pauseEvent is VMPauseExceptionEvent ||
+               isolate.pauseEvent is VMPauseInterruptedEvent) {
+      // If the isolate is paused for any other reason, assume the extension is
+      // already there.
+      _log('Isolate is paused mid-flight.');
+      await resumeLeniently();
+    } else if (isolate.pauseEvent is VMResumeEvent) {
+      _log('Isolate is not paused. Assuming application is ready.');
+    } else {
+      _log(
+        'Unknown pause event type ${isolate.pauseEvent.runtimeType}. '
+        'Assuming application is ready.'
+      );
+    }
+
+    // Invoked checkHealth and try to fix delays in the registration of Service
+    // extensions
+    Future<Health> checkHealth() async {
+      try {
+        // At this point the service extension must be installed. Verify it.
+        return await driver.checkHealth();
+      } on rpc.RpcException catch (e) {
+        if (e.code != error_code.METHOD_NOT_FOUND) {
+          rethrow;
+        }
+        _log('Check Health failed, try to wait for the service extensions to be registered.');
+        await enableIsolateStreams();
+        await waitForServiceExtension();
+        return driver.checkHealth();
+      }
+    }
+
+    final Health health = await checkHealth();
+    if (health.status != HealthStatus.ok) {
+      await client.close();
+      throw DriverError('Flutter application health check failed.');
+    }
+
+    _log('Connected to Flutter application.');
+    return driver;
   }
 
-  /// Sends [command] to the Flutter Driver extensions.
-  /// This must be implemented by subclass.
+  /// The unique ID of this driver instance.
+  final int _driverId;
+
+  /// Client connected to the Dart VM running the Flutter application
   ///
-  /// See also:
+  /// You can use [VMServiceClient] to check VM version, flags and get
+  /// notified when a new isolate has been instantiated. That could be
+  /// useful if your application spawns multiple isolates that you
+  /// would like to instrument.
+  final VMServiceClient serviceClient;
+
+  /// JSON-RPC client useful for sending raw JSON requests.
+  rpc.Peer _peer;
+
+  String _dartVmReconnectUrl;
+
+  Future<void> _restorePeerConnectionIfNeeded() async {
+    if (!_peer.isClosed || _dartVmReconnectUrl == null) {
+      return;
+    }
+    _log('Peer connection is closed! Trying to restore the connection...');
+    final String webSocketUrl = _getWebSocketUrl(_dartVmReconnectUrl);
+    final WebSocket ws = await WebSocket.connect(webSocketUrl);
+    ws.done.whenComplete(() => _checkCloseCode(ws));
+    _peer = rpc.Peer(
+        IOWebSocketChannel(ws).cast(),
+        onUnhandledError: _unhandledJsonRpcError,
+    )..listen();
+  }
+
+  /// The main isolate hosting the Flutter application.
   ///
-  ///  * [VMServiceFlutterDriver], which uses vmservice to implement.
-  ///  * [WebFlutterDriver], which uses webdriver to implement.
-  Future<Map<String, dynamic>> sendCommand(Command command);
+  /// If you used the [registerExtension] API to instrument your application,
+  /// you can use this [VMIsolate] to call these extension methods via
+  /// [invokeExtension].
+  final VMIsolate appIsolate;
+
+  /// Whether to print communication between host and app to `stdout`.
+  final bool _printCommunication;
+
+  /// Whether to log communication between host and app to `flutter_driver_commands.log`.
+  final bool _logCommunicationToFile;
+
+  Future<Map<String, dynamic>> _sendCommand(Command command) async {
+    Map<String, dynamic> response;
+    try {
+      final Map<String, String> serialized = command.serialize();
+      _logCommunication('>>> $serialized');
+      final Future<Map<String, dynamic>> future = appIsolate.invokeExtension(
+        _flutterExtensionMethodName,
+        serialized,
+      ).then<Map<String, dynamic>>((Object value) => value as Map<String, dynamic>);
+      response = await _warnIfSlow<Map<String, dynamic>>(
+        future: future,
+        timeout: command.timeout ?? kUnusuallyLongTimeout,
+        message: '${command.kind} message is taking a long time to complete...',
+      );
+      _logCommunication('<<< $response');
+    } catch (error, stackTrace) {
+      throw DriverError(
+        'Failed to fulfill ${command.runtimeType} due to remote error',
+        error,
+        stackTrace,
+      );
+    }
+    if (response['isError'] as bool)
+      throw DriverError('Error in Flutter application: ${response['response']}');
+    return response['response'] as Map<String, dynamic>;
+  }
+
+  void _logCommunication(String message) {
+    if (_printCommunication)
+      _log(message);
+    if (_logCommunicationToFile) {
+      final f.File file = fs.file(p.join(testOutputsDirectory, 'flutter_driver_commands_$_driverId.log'));
+      file.createSync(recursive: true); // no-op if file exists
+      file.writeAsStringSync('${DateTime.now()} $message\n', mode: f.FileMode.append, flush: true);
+    }
+  }
 
   /// Checks the status of the Flutter Driver extension.
   Future<Health> checkHealth({ Duration timeout }) async {
-    return Health.fromJson(await sendCommand(GetHealth(timeout: timeout)));
+    return Health.fromJson(await _sendCommand(GetHealth(timeout: timeout)));
   }
 
   /// Returns a dump of the render tree.
   Future<RenderTree> getRenderTree({ Duration timeout }) async {
-    return RenderTree.fromJson(await sendCommand(GetRenderTree(timeout: timeout)));
+    return RenderTree.fromJson(await _sendCommand(GetRenderTree(timeout: timeout)));
   }
 
   /// Returns a dump of the layer tree.
   Future<LayerTree> getLayerTree({ Duration timeout }) async {
-    return LayerTree.fromJson(await sendCommand(GetLayerTree(timeout: timeout)));
+    return LayerTree.fromJson(await _sendCommand(GetLayerTree(timeout: timeout)));
   }
 
   /// Taps at the center of the widget located by [finder].
   Future<void> tap(SerializableFinder finder, { Duration timeout }) async {
-    await sendCommand(Tap(finder, timeout: timeout));
+    await _sendCommand(Tap(finder, timeout: timeout));
   }
 
   /// Waits until [finder] locates the target.
   Future<void> waitFor(SerializableFinder finder, { Duration timeout }) async {
-    await sendCommand(WaitFor(finder, timeout: timeout));
+    await _sendCommand(WaitFor(finder, timeout: timeout));
   }
 
   /// Waits until [finder] can no longer locate the target.
   Future<void> waitForAbsent(SerializableFinder finder, { Duration timeout }) async {
-    await sendCommand(WaitForAbsent(finder, timeout: timeout));
+    await _sendCommand(WaitForAbsent(finder, timeout: timeout));
   }
 
   /// Waits until the given [waitCondition] is satisfied.
   Future<void> waitForCondition(SerializableWaitCondition waitCondition, {Duration timeout}) async {
-    await sendCommand(WaitForCondition(waitCondition, timeout: timeout));
+    await _sendCommand(WaitForCondition(waitCondition, timeout: timeout));
   }
 
   /// Waits until there are no more transient callbacks in the queue.
@@ -205,22 +506,20 @@ abstract class FlutterDriver {
   /// Use this method when you need to wait for the moment when the application
   /// becomes "stable", for example, prior to taking a [screenshot].
   Future<void> waitUntilNoTransientCallbacks({ Duration timeout }) async {
-    await sendCommand(WaitForCondition(const NoTransientCallbacks(), timeout: timeout));
+    await _sendCommand(WaitForCondition(const NoTransientCallbacks(), timeout: timeout));
   }
 
   /// Waits until the next [Window.onReportTimings] is called.
   ///
   /// Use this method to wait for the first frame to be rasterized during the
   /// app launch.
-  ///
-  /// Throws [UnimplementedError] on [WebFlutterDriver] instances.
   Future<void> waitUntilFirstFrameRasterized() async {
-    await sendCommand(const WaitForCondition(FirstFrameRasterized()));
+    await _sendCommand(const WaitForCondition(FirstFrameRasterized()));
   }
 
   Future<DriverOffset> _getOffset(SerializableFinder finder, OffsetType type, { Duration timeout }) async {
     final GetOffset command = GetOffset(finder, type, timeout: timeout);
-    final GetOffsetResult result = GetOffsetResult.fromJson(await sendCommand(command));
+    final GetOffsetResult result = GetOffsetResult.fromJson(await _sendCommand(command));
     return DriverOffset(result.dx, result.dy);
   }
 
@@ -291,7 +590,7 @@ abstract class FlutterDriver {
       bool includeProperties = true,
       Duration timeout,
   }) async {
-    return sendCommand(GetDiagnosticsTree(
+    return _sendCommand(GetDiagnosticsTree(
       finder,
       DiagnosticsType.renderObject,
       subtreeDepth: subtreeDepth,
@@ -324,7 +623,7 @@ abstract class FlutterDriver {
     bool includeProperties = true,
     Duration timeout,
   }) async {
-    return sendCommand(GetDiagnosticsTree(
+    return _sendCommand(GetDiagnosticsTree(
       finder,
       DiagnosticsType.widget,
       subtreeDepth: subtreeDepth,
@@ -347,7 +646,7 @@ abstract class FlutterDriver {
   /// The move events are generated at a given [frequency] in Hz (or events per
   /// second). It defaults to 60Hz.
   Future<void> scroll(SerializableFinder finder, double dx, double dy, Duration duration, { int frequency = 60, Duration timeout }) async {
-    await sendCommand(Scroll(finder, dx, dy, duration, frequency, timeout: timeout));
+    await _sendCommand(Scroll(finder, dx, dy, duration, frequency, timeout: timeout));
   }
 
   /// Scrolls the Scrollable ancestor of the widget located by [finder]
@@ -358,7 +657,7 @@ abstract class FlutterDriver {
   /// then this method may fail because [finder] doesn't actually exist.
   /// The [scrollUntilVisible] method can be used in this case.
   Future<void> scrollIntoView(SerializableFinder finder, { double alignment = 0.0, Duration timeout }) async {
-    await sendCommand(ScrollIntoView(finder, alignment: alignment, timeout: timeout));
+    await _sendCommand(ScrollIntoView(finder, alignment: alignment, timeout: timeout));
   }
 
   /// Repeatedly [scroll] the widget located by [scrollable] by [dxScroll] and
@@ -415,7 +714,7 @@ abstract class FlutterDriver {
 
   /// Returns the text in the `Text` widget located by [finder].
   Future<String> getText(SerializableFinder finder, { Duration timeout }) async {
-    return GetTextResult.fromJson(await sendCommand(GetText(finder, timeout: timeout))).text;
+    return GetTextResult.fromJson(await _sendCommand(GetText(finder, timeout: timeout))).text;
   }
 
   /// Enters `text` into the currently focused text input, such as the
@@ -451,7 +750,7 @@ abstract class FlutterDriver {
   /// });
   /// ```
   Future<void> enterText(String text, { Duration timeout }) async {
-    await sendCommand(EnterText(text, timeout: timeout));
+    await _sendCommand(EnterText(text, timeout: timeout));
   }
 
   /// Configures text entry emulation.
@@ -469,7 +768,7 @@ abstract class FlutterDriver {
   /// channel will be mocked out.
   Future<void> setTextEntryEmulation({ @required bool enabled, Duration timeout }) async {
     assert(enabled != null);
-    await sendCommand(SetTextEntryEmulation(enabled, timeout: timeout));
+    await _sendCommand(SetTextEntryEmulation(enabled, timeout: timeout));
   }
 
   /// Sends a string and returns a string.
@@ -479,7 +778,7 @@ abstract class FlutterDriver {
   /// callback in [enableFlutterDriverExtension] that can successfully handle
   /// these requests.
   Future<String> requestData(String message, { Duration timeout }) async {
-    return RequestDataResult.fromJson(await sendCommand(RequestData(message, timeout: timeout))).message;
+    return RequestDataResult.fromJson(await _sendCommand(RequestData(message, timeout: timeout))).message;
   }
 
   /// Turns semantics on or off in the Flutter app under test.
@@ -487,7 +786,7 @@ abstract class FlutterDriver {
   /// Returns true when the call actually changed the state from on to off or
   /// vice versa.
   Future<bool> setSemantics(bool enabled, { Duration timeout }) async {
-    final SetSemanticsResult result = SetSemanticsResult.fromJson(await sendCommand(SetSemantics(enabled, timeout: timeout)));
+    final SetSemanticsResult result = SetSemanticsResult.fromJson(await _sendCommand(SetSemantics(enabled, timeout: timeout)));
     return result.changedState;
   }
 
@@ -500,7 +799,7 @@ abstract class FlutterDriver {
   /// Semantics must be enabled to use this method, either using a platform
   /// specific shell command or [setSemantics].
   Future<int> getSemanticsId(SerializableFinder finder, { Duration timeout }) async {
-    final Map<String, dynamic> jsonResponse = await sendCommand(GetSemanticsId(finder, timeout: timeout));
+    final Map<String, dynamic> jsonResponse = await _sendCommand(GetSemanticsId(finder, timeout: timeout));
     final GetSemanticsIdResult result = GetSemanticsIdResult.fromJson(jsonResponse);
     return result.id;
   }
@@ -508,54 +807,58 @@ abstract class FlutterDriver {
   /// Take a screenshot.
   ///
   /// The image will be returned as a PNG.
-  ///
-  ///  HACK: There will be a 2-second artificial delay before screenshotting,
-  ///        the delay here is to deal with a race between the driver script and
-  ///        the GPU thread. The issue is that driver API synchronizes with the
-  ///        framework based on transient callbacks, which are out of sync with
-  ///        the GPU thread. Here's the timeline of events in ASCII art:
-  ///
-  ///        -------------------------------------------------------------------
-  ///        Without this delay:
-  ///        -------------------------------------------------------------------
-  ///        UI    : <-- build -->
-  ///        GPU   :               <-- rasterize -->
-  ///        Gap   :              | random |
-  ///        Driver:                        <-- screenshot -->
-  ///
-  ///        In the diagram above, the gap is the time between the last driver
-  ///        action taken, such as a `tap()`, and the subsequent call to
-  ///        `screenshot()`. The gap is random because it is determined by the
-  ///        unpredictable network communication between the driver process and
-  ///        the application. If this gap is too short, which it typically will
-  ///        be, the screenshot is taken before the GPU thread is done
-  ///        rasterizing the frame, so the screenshot of the previous frame is
-  ///        taken, which is wrong.
-  ///
-  ///        -------------------------------------------------------------------
-  ///        With this delay, if we're lucky:
-  ///        -------------------------------------------------------------------
-  ///        UI    : <-- build -->
-  ///        GPU   :               <-- rasterize -->
-  ///        Gap   :              |    2 seconds or more   |
-  ///        Driver:                                        <-- screenshot -->
-  ///
-  ///        The two-second gap should be long enough for the GPU thread to
-  ///        finish rasterizing the frame, but not longer than necessary to keep
-  ///        driver tests as fast a possible.
-  ///
-  ///        -------------------------------------------------------------------
-  ///        With this delay, if we're not lucky:
-  ///        -------------------------------------------------------------------
-  ///        UI    : <-- build -->
-  ///        GPU   :               <-- rasterize randomly slow today -->
-  ///        Gap   :              |    2 seconds or more   |
-  ///        Driver:                                        <-- screenshot -->
-  ///
-  ///        In practice, sometimes the device gets really busy for a while and
-  ///        even two seconds isn't enough, which means that this is still racy
-  ///        and a source of flakes.
-  Future<List<int>> screenshot();
+  Future<List<int>> screenshot() async {
+    // HACK: this artificial delay here is to deal with a race between the
+    //       driver script and the GPU thread. The issue is that driver API
+    //       synchronizes with the framework based on transient callbacks, which
+    //       are out of sync with the GPU thread. Here's the timeline of events
+    //       in ASCII art:
+    //
+    //       -------------------------------------------------------------------
+    //       Without this delay:
+    //       -------------------------------------------------------------------
+    //       UI    : <-- build -->
+    //       GPU   :               <-- rasterize -->
+    //       Gap   :              | random |
+    //       Driver:                        <-- screenshot -->
+    //
+    //       In the diagram above, the gap is the time between the last driver
+    //       action taken, such as a `tap()`, and the subsequent call to
+    //       `screenshot()`. The gap is random because it is determined by the
+    //       unpredictable network communication between the driver process and
+    //       the application. If this gap is too short, which it typically will
+    //       be, the screenshot is taken before the GPU thread is done
+    //       rasterizing the frame, so the screenshot of the previous frame is
+    //       taken, which is wrong.
+    //
+    //       -------------------------------------------------------------------
+    //       With this delay, if we're lucky:
+    //       -------------------------------------------------------------------
+    //       UI    : <-- build -->
+    //       GPU   :               <-- rasterize -->
+    //       Gap   :              |    2 seconds or more   |
+    //       Driver:                                        <-- screenshot -->
+    //
+    //       The two-second gap should be long enough for the GPU thread to
+    //       finish rasterizing the frame, but not longer than necessary to keep
+    //       driver tests as fast a possible.
+    //
+    //       -------------------------------------------------------------------
+    //       With this delay, if we're not lucky:
+    //       -------------------------------------------------------------------
+    //       UI    : <-- build -->
+    //       GPU   :               <-- rasterize randomly slow today -->
+    //       Gap   :              |    2 seconds or more   |
+    //       Driver:                                        <-- screenshot -->
+    //
+    //       In practice, sometimes the device gets really busy for a while and
+    //       even two seconds isn't enough, which means that this is still racy
+    //       and a source of flakes.
+    await Future<void>.delayed(const Duration(seconds: 2));
+
+    final Map<String, dynamic> result = await _peer.sendRequest('_flutter.screenshot') as Map<String, dynamic>;
+    return base64.decode(result['screenshot'] as String);
+  }
 
   /// Returns the Flags set in the Dart VM as JSON.
   ///
@@ -576,32 +879,76 @@ abstract class FlutterDriver {
   ///     ]
   ///
   /// [getFlagList]: https://github.com/dart-lang/sdk/blob/master/runtime/vm/service/service.md#getflaglist
-  ///
-  /// Throws [UnimplementedError] on [WebFlutterDriver] instances.
-  Future<List<Map<String, dynamic>>> getVmFlags();
+  Future<List<Map<String, dynamic>>> getVmFlags() async {
+    await _restorePeerConnectionIfNeeded();
+    final Map<String, dynamic> result = await _peer.sendRequest('getFlagList') as Map<String, dynamic>;
+    return result != null
+        ? (result['flags'] as List<dynamic>).cast<Map<String,dynamic>>()
+        : const <Map<String, dynamic>>[];
+  }
 
   /// Starts recording performance traces.
   ///
   /// The `timeout` argument causes a warning to be displayed to the user if the
   /// operation exceeds the specified timeout; it does not actually cancel the
   /// operation.
-  ///
-  /// For [WebFlutterDriver], this is only supported for Chrome.
   Future<void> startTracing({
-    List<TimelineStream> streams = const <TimelineStream>[TimelineStream.all],
+    List<TimelineStream> streams = _defaultStreams,
     Duration timeout = kUnusuallyLongTimeout,
-  });
+  }) async {
+    assert(streams != null && streams.isNotEmpty);
+    assert(timeout != null);
+    try {
+      await _warnIfSlow<void>(
+        future: _peer.sendRequest(_setVMTimelineFlagsMethodName, <String, String>{
+          'recordedStreams': _timelineStreamsToString(streams),
+        }),
+        timeout: timeout,
+        message: 'VM is taking an unusually long time to respond to being told to start tracing...',
+      );
+    } catch (error, stackTrace) {
+      throw DriverError(
+        'Failed to start tracing due to remote error',
+        error,
+        stackTrace,
+      );
+    }
+  }
 
   /// Stops recording performance traces and downloads the timeline.
   ///
   /// The `timeout` argument causes a warning to be displayed to the user if the
   /// operation exceeds the specified timeout; it does not actually cancel the
   /// operation.
-  ///
-  /// For [WebFlutterDriver], this is only supported for Chrome.
   Future<Timeline> stopTracingAndDownloadTimeline({
     Duration timeout = kUnusuallyLongTimeout,
-  });
+  }) async {
+    assert(timeout != null);
+    try {
+      await _warnIfSlow<void>(
+        future: _peer.sendRequest(_setVMTimelineFlagsMethodName, <String, String>{'recordedStreams': '[]'}),
+        timeout: timeout,
+        message: 'VM is taking an unusually long time to respond to being told to stop tracing...',
+      );
+      return Timeline.fromJson(await _peer.sendRequest(_getVMTimelineMethodName) as Map<String, dynamic>);
+    } catch (error, stackTrace) {
+      throw DriverError(
+        'Failed to stop tracing due to remote error',
+        error,
+        stackTrace,
+      );
+    }
+  }
+
+  Future<bool> _isPrecompiledMode() async {
+    final List<Map<String, dynamic>> flags = await getVmFlags();
+    for(Map<String, dynamic> flag in flags) {
+      if (flag['name'] == 'precompiled_mode') {
+        return flag['valueAsString'] == 'true';
+      }
+    }
+    return false;
+  }
 
   /// Runs [action] and outputs a performance trace for it.
   ///
@@ -620,24 +967,46 @@ abstract class FlutterDriver {
   ///
   /// If this is run in debug mode, a warning message will be printed to suggest
   /// running the benchmark in profile mode instead.
-  ///
-  /// For [WebFlutterDriver], this is only supported for Chrome.
   Future<Timeline> traceAction(
     Future<dynamic> action(), {
-    List<TimelineStream> streams = const <TimelineStream>[TimelineStream.all],
+    List<TimelineStream> streams = _defaultStreams,
     bool retainPriorEvents = false,
-  });
+  }) async {
+    if (!retainPriorEvents) {
+      await clearTimeline();
+    }
+    await startTracing(streams: streams);
+    await action();
+
+    if (!(await _isPrecompiledMode())) {
+      _log(_kDebugWarning);
+    }
+    return stopTracingAndDownloadTimeline();
+  }
 
   /// Clears all timeline events recorded up until now.
   ///
   /// The `timeout` argument causes a warning to be displayed to the user if the
   /// operation exceeds the specified timeout; it does not actually cancel the
   /// operation.
-  ///
-  /// For [WebFlutterDriver], this is only supported for Chrome.
   Future<void> clearTimeline({
     Duration timeout = kUnusuallyLongTimeout,
-  });
+  }) async {
+    assert(timeout != null);
+    try {
+      await _warnIfSlow<void>(
+        future: _peer.sendRequest(_clearVMTimelineMethodName, <String, String>{}),
+        timeout: timeout,
+        message: 'VM is taking an unusually long time to respond to being told to clear its timeline buffer...',
+      );
+    } catch (error, stackTrace) {
+      throw DriverError(
+        'Failed to clear event timeline due to remote error',
+        error,
+        stackTrace,
+      );
+    }
+  }
 
   /// [action] will be executed with the frame sync mechanism disabled.
   ///
@@ -656,25 +1025,155 @@ abstract class FlutterDriver {
   /// ensure that no action is performed while the app is undergoing a
   /// transition to avoid flakiness.
   Future<T> runUnsynchronized<T>(Future<T> action(), { Duration timeout }) async {
-    await sendCommand(SetFrameSync(false, timeout: timeout));
+    await _sendCommand(SetFrameSync(false, timeout: timeout));
     T result;
     try {
       result = await action();
     } finally {
-      await sendCommand(SetFrameSync(true, timeout: timeout));
+      await _sendCommand(SetFrameSync(true, timeout: timeout));
     }
     return result;
   }
 
   /// Force a garbage collection run in the VM.
-  ///
-  /// Throws [UnimplementedError] on [WebFlutterDriver] instances.
-  Future<void> forceGC();
+  Future<void> forceGC() async {
+    try {
+      await _peer
+          .sendRequest(_collectAllGarbageMethodName, <String, String>{
+            'isolateId': 'isolates/${appIsolate.numberAsString}',
+          });
+    } catch (error, stackTrace) {
+      throw DriverError(
+        'Failed to force a GC due to remote error',
+        error,
+        stackTrace,
+      );
+    }
+  }
 
   /// Closes the underlying connection to the VM service.
   ///
   /// Returns a [Future] that fires once the connection has been closed.
-  Future<void> close();
+  Future<void> close() async {
+    // Don't leak vm_service_client-specific objects, if any
+    await serviceClient.close();
+    await _peer.close();
+  }
+}
+
+/// Encapsulates connection information to an instance of a Flutter application.
+@visibleForTesting
+class VMServiceClientConnection {
+  /// Creates an instance of this class given a [client] and a [peer].
+  VMServiceClientConnection(this.client, this.peer);
+
+  /// Use this for structured access to the VM service's public APIs.
+  final VMServiceClient client;
+
+  /// Use this to make arbitrary raw JSON-RPC calls.
+  ///
+  /// This object allows reaching into private VM service APIs. Use with
+  /// caution.
+  final rpc.Peer peer;
+}
+
+/// A function that connects to a Dart VM service given the [url].
+typedef VMServiceConnectFunction = Future<VMServiceClientConnection> Function(String url);
+
+/// The connection function used by [FlutterDriver.connect].
+///
+/// Overwrite this function if you require a custom method for connecting to
+/// the VM service.
+VMServiceConnectFunction vmServiceConnectFunction = _waitAndConnect;
+
+/// Restores [vmServiceConnectFunction] to its default value.
+void restoreVmServiceConnectFunction() {
+  vmServiceConnectFunction = _waitAndConnect;
+}
+
+/// The JSON RPC 2 spec says that a notification from a client must not respond
+/// to the client. It's possible the client sent a notification as a "ping", but
+/// the service isn't set up yet to respond.
+///
+/// For example, if the client sends a notification message to the server for
+/// 'streamNotify', but the server has not finished loading, it will throw an
+/// exception. Since the message is a notification, the server follows the
+/// specification and does not send a response back, but is left with an
+/// unhandled exception. That exception is safe for us to ignore - the client
+/// is signaling that it will try again later if it doesn't get what it wants
+/// here by sending a notification.
+// This may be ignoring too many exceptions. It would be best to rewrite
+// the client code to not use notifications so that it gets error replies back
+// and can decide what to do from there.
+// TODO(dnfield): https://github.com/flutter/flutter/issues/31813
+bool _ignoreRpcError(dynamic error) {
+  if (error is rpc.RpcException) {
+    final rpc.RpcException exception = error;
+    return exception.data == null || exception.data['id'] == null;
+  } else if (error is String && error.startsWith('JSON-RPC error -32601')) {
+    return true;
+  }
+  return false;
+}
+
+void _unhandledJsonRpcError(dynamic error, dynamic stack) {
+  if (_ignoreRpcError(error)) {
+    return;
+  }
+  _log('Unhandled RPC error:\n$error\n$stack');
+  // TODO(dnfield): https://github.com/flutter/flutter/issues/31813
+  // assert(false);
+}
+
+String _getWebSocketUrl(String url) {
+  Uri uri = Uri.parse(url);
+  final List<String> pathSegments = <String>[
+    // If there's an authentication code (default), we need to add it to our path.
+    if (uri.pathSegments.isNotEmpty) uri.pathSegments.first,
+    'ws',
+  ];
+  if (uri.scheme == 'http')
+    uri = uri.replace(scheme: 'ws', pathSegments: pathSegments);
+  return uri.toString();
+}
+
+void _checkCloseCode(WebSocket ws) {
+  if (ws.closeCode != 1000 && ws.closeCode != null) {
+    _log('$ws is closed with an unexpected code ${ws.closeCode}');
+  }
+}
+
+/// Waits for a real Dart VM service to become available, then connects using
+/// the [VMServiceClient].
+Future<VMServiceClientConnection> _waitAndConnect(String url) async {
+  final String webSocketUrl = _getWebSocketUrl(url);
+  int attempts = 0;
+  while (true) {
+    WebSocket ws1;
+    WebSocket ws2;
+    try {
+      ws1 = await WebSocket.connect(webSocketUrl);
+      ws2 = await WebSocket.connect(webSocketUrl);
+
+      ws1.done.whenComplete(() => _checkCloseCode(ws1));
+      ws2.done.whenComplete(() => _checkCloseCode(ws2));
+
+      return VMServiceClientConnection(
+        VMServiceClient(IOWebSocketChannel(ws1).cast()),
+        rpc.Peer(
+            IOWebSocketChannel(ws2).cast(),
+            onUnhandledError: _unhandledJsonRpcError,
+        )..listen(),
+      );
+    } catch (e) {
+      await ws1?.close();
+      await ws2?.close();
+      if (attempts > 5)
+        _log('It is taking an unusually long time to connect to the VM...');
+      attempts += 1;
+      await Future<void>.delayed(_kPauseBetweenReconnectAttempts);
+    }
+  }
 }
 
 /// Provides convenient accessors to frequently used finders.
