+++ packages/flutter_driver/pubspec.yaml
@@ -14,7 +14,6 @@ dependencies:
   path: 1.6.4
   web_socket_channel: 1.1.0
   vm_service_client: 0.2.6+2
-  webdriver: 2.1.1
   flutter:
     sdk: flutter
   flutter_test:
@@ -42,7 +41,6 @@ dependencies:
   stack_trace: 1.9.3 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
   stream_channel: 2.0.0 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
   string_scanner: 1.0.5 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
-  sync_http: 0.1.4 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
   term_glyph: 1.1.0 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
   test_api: 0.2.11 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
   typed_data: 1.1.6 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
@@ -53,4 +51,4 @@ dev_dependencies:
   mockito: 4.1.1
   quiver: 2.0.5
 
-# PUBSPEC CHECKSUM: 1daa
+# PUBSPEC CHECKSUM: a765
