+++ packages/flutter_driver/test/flutter_driver_test.dart
@@ -3,7 +3,6 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:convert';
 
 import 'package:flutter_driver/src/common/error.dart';
 import 'package:flutter_driver/src/common/health.dart';
@@ -21,8 +20,6 @@ import 'common.dart';
 /// Magical timeout value that's different from the default.
 const Duration _kTestTimeout = Duration(milliseconds: 1234);
 const String _kSerializedTestTimeout = '1234';
-const String _kWebScriptPrefix = 'window.\$flutterDriver(\'';
-const String _kWebScriptSuffix = '\')';
 
 void main() {
   final List<String> log = <String>[];
@@ -30,7 +27,7 @@ void main() {
     log.add('$source: $message');
   };
 
-  group('VMServiceFlutterDriver.connect', () {
+  group('FlutterDriver.connect', () {
     MockVMServiceClient mockClient;
     MockVM mockVM;
     MockIsolate mockIsolate;
@@ -118,17 +115,17 @@ void main() {
     });
   });
 
-  group('VMServiceFlutterDriver', () {
+  group('FlutterDriver', () {
     MockVMServiceClient mockClient;
     MockPeer mockPeer;
     MockIsolate mockIsolate;
-    VMServiceFlutterDriver driver;
+    FlutterDriver driver;
 
     setUp(() {
       mockClient = MockVMServiceClient();
       mockPeer = MockPeer();
       mockIsolate = MockIsolate();
-      driver = VMServiceFlutterDriver.connectedTo(mockClient, mockPeer, mockIsolate);
+      driver = FlutterDriver.connectedTo(mockClient, mockPeer, mockIsolate);
     });
 
     test('checks the health of the driver extension', () async {
@@ -618,7 +615,7 @@ void main() {
           expect(log, <String>[]);
           time.elapse(kUnusuallyLongTimeout);
         });
-        expect(log, <String>['VMServiceFlutterDriver: waitFor message is taking a long time to complete...']);
+        expect(log, <String>['FlutterDriver: waitFor message is taking a long time to complete...']);
       });
 
       test('local custom timeout', () async {
@@ -633,7 +630,7 @@ void main() {
           expect(log, <String>[]);
           time.elapse(customTimeout);
         });
-        expect(log, <String>['VMServiceFlutterDriver: waitFor message is taking a long time to complete...']);
+        expect(log, <String>['FlutterDriver: waitFor message is taking a long time to complete...']);
       });
 
       test('remote error', () async {
@@ -653,17 +650,17 @@ void main() {
     });
   });
 
-  group('VMServiceFlutterDriver with custom timeout', () {
+  group('FlutterDriver with custom timeout', () {
     MockVMServiceClient mockClient;
     MockPeer mockPeer;
     MockIsolate mockIsolate;
-    VMServiceFlutterDriver driver;
+    FlutterDriver driver;
 
     setUp(() {
       mockClient = MockVMServiceClient();
       mockPeer = MockPeer();
       mockIsolate = MockIsolate();
-      driver = VMServiceFlutterDriver.connectedTo(mockClient, mockPeer, mockIsolate);
+      driver = FlutterDriver.connectedTo(mockClient, mockPeer, mockIsolate);
     });
 
     test('GetHealth has no default timeout', () async {
@@ -687,377 +684,6 @@ void main() {
       await driver.checkHealth(timeout: _kTestTimeout);
     });
   });
-
-  group('WebFlutterDriver', () {
-    MockFlutterWebConnection mockConnection;
-    WebFlutterDriver driver;
-
-    setUp(() {
-      mockConnection = MockFlutterWebConnection();
-      driver = WebFlutterDriver.connectedTo(mockConnection, Browser.chrome);
-    });
-
-    test('closes connection', () async {
-      when(mockConnection.close()).thenAnswer((Invocation invocation) => Future<dynamic>.value(null));
-      await driver.close();
-    });
-
-    group('ByValueKey', () {
-      test('restricts value types', () async {
-        expect(() => find.byValueKey(null),
-            throwsA(isInstanceOf<DriverError>()));
-      });
-
-      test('finds by ValueKey', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'tap',
-            'timeout': _kSerializedTestTimeout,
-            'finderType': 'ByValueKey',
-            'keyValueString': 'foo',
-            'keyValueType': 'String',
-          });
-          return jsonEncode(await makeMockResponse(<String, dynamic>{}));
-        });
-        await driver.tap(find.byValueKey('foo'), timeout: _kTestTimeout);
-      });
-    });
-
-    group('BySemanticsLabel', () {
-      test('finds by Semantic label using String', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'tap',
-            'timeout': _kSerializedTestTimeout,
-            'finderType': 'BySemanticsLabel',
-            'label': 'foo',
-          });
-          return jsonEncode(await makeMockResponse(<String, dynamic>{}));
-        });
-        await driver.tap(find.bySemanticsLabel('foo'), timeout: _kTestTimeout);
-      });
-
-      test('finds by Semantic label using RegExp', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'tap',
-            'timeout': _kSerializedTestTimeout,
-            'finderType': 'BySemanticsLabel',
-            'label': '^foo',
-            'isRegExp': 'true',
-          });
-          return jsonEncode(await makeMockResponse(<String, dynamic>{}));
-        });
-        await driver.tap(find.bySemanticsLabel(RegExp('^foo')), timeout: _kTestTimeout);
-      });
-    });
-
-    group('tap', () {
-      test('requires a target reference', () async {
-        expect(driver.tap(null), throwsA(isInstanceOf<DriverError>()));
-      });
-
-      test('sends the tap command', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'tap',
-            'timeout': _kSerializedTestTimeout,
-            'finderType': 'ByText',
-            'text': 'foo',
-          });
-          return jsonEncode(await makeMockResponse(<String, dynamic>{}));
-        });
-        await driver.tap(find.text('foo'), timeout: _kTestTimeout);
-      });
-    });
-
-    group('getText', () {
-      test('requires a target reference', () async {
-        expect(driver.getText(null), throwsA(isInstanceOf<DriverError>()));
-      });
-
-      test('sends the getText command', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'get_text',
-            'timeout': _kSerializedTestTimeout,
-            'finderType': 'ByValueKey',
-            'keyValueString': '123',
-            'keyValueType': 'int',
-          });
-          return jsonEncode(await makeMockResponse(<String, String>{
-            'text': 'hello',
-          }));
-        });
-        final String result = await driver.getText(find.byValueKey(123), timeout: _kTestTimeout);
-        expect(result, 'hello');
-      });
-    });
-
-    group('waitFor', () {
-      test('requires a target reference', () async {
-        expect(driver.waitFor(null), throwsA(isInstanceOf<DriverError>()));
-      });
-
-      test('sends the waitFor command', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'waitFor',
-            'finderType': 'ByTooltipMessage',
-            'text': 'foo',
-            'timeout': _kSerializedTestTimeout,
-          });
-          return jsonEncode(await makeMockResponse(<String, dynamic>{}));
-        });
-        await driver.waitFor(find.byTooltip('foo'), timeout: _kTestTimeout);
-      });
-    });
-
-    group('waitForCondition', () {
-      test('sends the wait for NoPendingFrameCondition command', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'waitForCondition',
-            'timeout': _kSerializedTestTimeout,
-            'conditionName': 'NoPendingFrameCondition',
-          });
-          return jsonEncode(await makeMockResponse(<String, dynamic>{}));
-        });
-        await driver.waitForCondition(const NoPendingFrame(), timeout: _kTestTimeout);
-      });
-
-      test('sends the wait for NoPendingPlatformMessages command', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'waitForCondition',
-            'timeout': _kSerializedTestTimeout,
-            'conditionName': 'NoPendingPlatformMessagesCondition',
-          });
-          return jsonEncode(await makeMockResponse(<String, dynamic>{}));
-        });
-        await driver.waitForCondition(const NoPendingPlatformMessages(), timeout: _kTestTimeout);
-      });
-
-      test('sends the waitForCondition of combined conditions command', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'waitForCondition',
-            'timeout': _kSerializedTestTimeout,
-            'conditionName': 'CombinedCondition',
-            'conditions': '[{"conditionName":"NoPendingFrameCondition"},{"conditionName":"NoTransientCallbacksCondition"}]',
-          });
-          return jsonEncode(await makeMockResponse(<String, dynamic>{}));
-        });
-        const SerializableWaitCondition combinedCondition =
-        CombinedCondition(<SerializableWaitCondition>[NoPendingFrame(), NoTransientCallbacks()]);
-        await driver.waitForCondition(combinedCondition, timeout: _kTestTimeout);
-      });
-    });
-
-    group('waitUntilNoTransientCallbacks', () {
-      test('sends the waitUntilNoTransientCallbacks command', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'waitForCondition',
-            'timeout': _kSerializedTestTimeout,
-            'conditionName': 'NoTransientCallbacksCondition',
-          });
-          return jsonEncode(await makeMockResponse(<String, dynamic>{}));
-        });
-        await driver.waitUntilNoTransientCallbacks(timeout: _kTestTimeout);
-      });
-    });
-
-    group('getOffset', () {
-      test('requires a target reference', () async {
-        expect(driver.getCenter(null), throwsA(isInstanceOf<DriverError>()));
-        expect(driver.getTopLeft(null), throwsA(isInstanceOf<DriverError>()));
-        expect(driver.getTopRight(null), throwsA(isInstanceOf<DriverError>()));
-        expect(driver.getBottomLeft(null), throwsA(isInstanceOf<DriverError>()));
-        expect(driver.getBottomRight(null), throwsA(isInstanceOf<DriverError>()));
-      });
-
-      test('sends the getCenter command', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'get_offset',
-            'offsetType': 'center',
-            'timeout': _kSerializedTestTimeout,
-            'finderType': 'ByValueKey',
-            'keyValueString': '123',
-            'keyValueType': 'int',
-          });
-          return jsonEncode(await makeMockResponse(<String, double>{
-            'dx': 11,
-            'dy': 12,
-          }));
-        });
-        final DriverOffset result = await driver.getCenter(find.byValueKey(123), timeout: _kTestTimeout);
-        expect(result, const DriverOffset(11, 12));
-      });
-
-      test('sends the getTopLeft command', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'get_offset',
-            'offsetType': 'topLeft',
-            'timeout': _kSerializedTestTimeout,
-            'finderType': 'ByValueKey',
-            'keyValueString': '123',
-            'keyValueType': 'int',
-          });
-          return jsonEncode(await makeMockResponse(<String, double>{
-            'dx': 11,
-            'dy': 12,
-          }));
-        });
-        final DriverOffset result = await driver.getTopLeft(find.byValueKey(123), timeout: _kTestTimeout);
-        expect(result, const DriverOffset(11, 12));
-      });
-
-      test('sends the getTopRight command', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'get_offset',
-            'offsetType': 'topRight',
-            'timeout': _kSerializedTestTimeout,
-            'finderType': 'ByValueKey',
-            'keyValueString': '123',
-            'keyValueType': 'int',
-          });
-          return jsonEncode(await makeMockResponse(<String, double>{
-            'dx': 11,
-            'dy': 12,
-          }));
-        });
-        final DriverOffset result = await driver.getTopRight(find.byValueKey(123), timeout: _kTestTimeout);
-        expect(result, const DriverOffset(11, 12));
-      });
-
-      test('sends the getBottomLeft command', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'get_offset',
-            'offsetType': 'bottomLeft',
-            'timeout': _kSerializedTestTimeout,
-            'finderType': 'ByValueKey',
-            'keyValueString': '123',
-            'keyValueType': 'int',
-          });
-          return jsonEncode(await makeMockResponse(<String, double>{
-            'dx': 11,
-            'dy': 12,
-          }));
-        });
-        final DriverOffset result = await driver.getBottomLeft(find.byValueKey(123), timeout: _kTestTimeout);
-        expect(result, const DriverOffset(11, 12));
-      });
-
-      test('sends the getBottomRight command', () async {
-        when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-          final String script = _checkAndEncode(i.positionalArguments[0]);
-          expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-            'command': 'get_offset',
-            'offsetType': 'bottomRight',
-            'timeout': _kSerializedTestTimeout,
-            'finderType': 'ByValueKey',
-            'keyValueString': '123',
-            'keyValueType': 'int',
-          });
-          return jsonEncode(await makeMockResponse(<String, double>{
-            'dx': 11,
-            'dy': 12,
-          }));
-        });
-        final DriverOffset result = await driver.getBottomRight(find.byValueKey(123), timeout: _kTestTimeout);
-        expect(result, const DriverOffset(11, 12));
-      });
-    });
-
-    test('checks the health of the driver extension', () async {
-      when(mockConnection.sendCommand(any, any)).thenAnswer((Invocation i) async {
-        final String script = _checkAndEncode(i.positionalArguments[0]);
-        expect(Map<String, String>.from(jsonDecode(script) as Map<String, dynamic>), <String, String>{
-          'command': 'get_health',
-        });
-        return jsonEncode(await makeMockResponse(<String, dynamic>{'status': 'ok'}));
-      });
-      await driver.checkHealth();
-    });
-
-    group('clearTimeline', () {
-      test('clears timeline', () async {
-        await driver.startTracing();
-        expect(driver.startTime, isNotNull);
-        await driver.clearTimeline();
-        expect(driver.startTime, isNull);
-      });
-    });
-
-    group('WebFlutterDriver Unimplemented error', () {
-      test('forceGC', () async {
-        expect(driver.forceGC(),
-            throwsA(isInstanceOf<UnimplementedError>()));
-      });
-
-      test('getVmFlags', () async {
-        expect(driver.getVmFlags(),
-            throwsA(isInstanceOf<UnimplementedError>()));
-      });
-
-      test('waitUntilFirstFrameRasterized', () async {
-        expect(driver.waitUntilFirstFrameRasterized(),
-            throwsA(isInstanceOf<UnimplementedError>()));
-      });
-    });
-  });
-
-  group('WebFlutterDriver with non-chrome browser', () {
-    MockFlutterWebConnection mockConnection;
-    WebFlutterDriver driver;
-
-    setUp(() {
-      mockConnection = MockFlutterWebConnection();
-      driver = WebFlutterDriver.connectedTo(mockConnection, Browser.edge);
-    });
-
-    test('tracing', () async {
-      expect(driver.traceAction(() async { return Future<dynamic>.value(); }),
-          throwsA(isInstanceOf<UnimplementedError>()));
-      expect(driver.startTracing(),
-          throwsA(isInstanceOf<UnimplementedError>()));
-      expect(driver.stopTracingAndDownloadTimeline(),
-          throwsA(isInstanceOf<UnimplementedError>()));
-      expect(driver.clearTimeline(),
-          throwsA(isInstanceOf<UnimplementedError>()));
-    });
-  });
-}
-
-/// This function will verify the format of the script
-/// and return the actual script.
-/// script will be in the following format:
-//   window.flutterDriver('[actual script]')
-String _checkAndEncode(dynamic script) {
-  expect(script is String, isTrue);
-  expect(script.startsWith(_kWebScriptPrefix), isTrue);
-  expect(script.endsWith(_kWebScriptSuffix), isTrue);
-  // Strip prefix and suffix
-  return script.substring(_kWebScriptPrefix.length, script.length - 2) as String;
 }
 
 Future<Map<String, dynamic>> makeMockResponse(
@@ -1081,8 +707,6 @@ class MockVMPauseStartEvent extends Mock implements VMPauseStartEvent { }
 class MockVMPauseBreakpointEvent extends Mock implements VMPauseBreakpointEvent { }
 
 class MockVMResumeEvent extends Mock implements VMResumeEvent { }
-
-class MockFlutterWebConnection extends Mock implements FlutterWebConnection { }
 
 class MockPeer extends Mock implements rpc.Peer {
   @override
