--- packages/flutter_driver/test/src/real_tests/io_extension_test.dart
@@ -1,24 +1,0 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import 'package:flutter_driver/src/extension/io_extension.dart';
-
-import '../../common.dart';
-
-void main() {
-  group('test io_extension',() {
-    Future<Map<String, dynamic>> Function(Map<String, String>) call;
-
-    setUp(() {
-      call = (Map<String, String> args) async {
-        return Future<Map<String, dynamic>>.value(args);
-      };
-    });
-
-    test('io_extension should throw exception', () {
-      expect(() => registerWebServiceExtension(call),
-          throwsA(isInstanceOf<UnsupportedError>()));
-    });
-  });
-}
