+++ packages/flutter_driver/test/src/timeline_summary_test.dart
@@ -9,7 +9,7 @@ import 'package:flutter_driver/flutter_driver.dart';
 import 'package:flutter_driver/src/driver/common.dart';
 import 'package:path/path.dart' as path;
 
-import '../../common.dart';
+import '../common.dart';
 
 void main() {
   group('TimelineSummary', () {
