+++ packages/flutter_driver/test/src/timeline_test.dart
@@ -4,7 +4,7 @@
 
 import 'package:flutter_driver/src/driver/timeline.dart';
 
-import '../../common.dart';
+import '../common.dart';
 
 void main() {
   group('Timeline', () {
