+++ packages/flutter_tools/pubspec.yaml
@@ -11,9 +11,9 @@ dependencies:
   # To update these, use "flutter update-packages --force-upgrade".
   archive: 2.0.11
   args: 1.5.2
-  dwds: 0.8.5
+  dwds: 0.8.1
   completion: 0.2.1+1
-  coverage: 0.13.3+3
+  coverage: 0.13.3+1
   crypto: 2.1.3
   file: 5.1.0
   http: 0.12.0+2
@@ -46,7 +46,7 @@ dependencies:
   test_core: 0.2.15
 
   # Code generation dependencies
-  build_runner_core: 4.3.0
+  build_runner_core: 4.2.0
   dart_style: 1.3.3
   code_builder: 3.2.1
   build: 1.2.2
@@ -118,7 +118,7 @@ dev_dependencies:
   test: 1.9.4
   build_runner: 1.7.2
   build_vm_compilers: 1.0.4
-  build_test: 0.10.11
+  build_test: 0.10.10
 
   multi_server_socket: 1.0.2 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
   node_preamble: 1.4.8 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
@@ -128,4 +128,4 @@ dartdoc:
   # Exclude this package from the hosted API docs.
   nodoc: true
 
-# PUBSPEC CHECKSUM: 6e90
+# PUBSPEC CHECKSUM: de88
