+++ packages/flutter_tools/test/commands.shard/hermetic/drive_test.dart
@@ -46,7 +46,7 @@ void main() {
       appStarter = (DriveCommand command) {
         throw 'Unexpected call to appStarter';
       };
-      testRunner = (List<String> testArgs, Map<String, String> environment) {
+      testRunner = (List<String> testArgs, String observatoryUri) {
         throw 'Unexpected call to testRunner';
       };
       appStopper = (DriveCommand command) {
@@ -171,16 +171,8 @@ void main() {
       appStarter = expectAsync1((DriveCommand command) async {
         return LaunchResult.succeeded();
       });
-      testRunner = expectAsync2((List<String> testArgs, Map<String, String> environment) async {
+      testRunner = expectAsync2((List<String> testArgs, String observatoryUri) async {
         expect(testArgs, <String>[testFile]);
-        // VM_SERVICE_URL is not set by drive command arguments
-        expect(environment, <String, String>{
-          'VM_SERVICE_URL': 'null',
-          'SELENIUM_PORT': '4567',
-          'BROWSER_NAME': 'firefox',
-          'BROWSER_DIMENSION': '1024,768',
-          'HEADLESS': 'false',
-        });
         return null;
       });
       appStopper = expectAsync1((DriveCommand command) async {
@@ -195,10 +187,6 @@ void main() {
         'drive',
         '--target=$testApp',
         '--no-pub',
-        '--no-headless',
-        '--driver-port=4567',
-        '--browser-name=firefox',
-        '--browser-dimension=1024,768',
       ];
       await createTestCommandRunner(command).run(args);
       expect(testLogger.errorText, isEmpty);
@@ -216,7 +204,7 @@ void main() {
       appStarter = expectAsync1((DriveCommand command) async {
         return LaunchResult.succeeded();
       });
-      testRunner = (List<String> testArgs, Map<String, String> environment) async {
+      testRunner = (List<String> testArgs, String observatoryUri) async {
         throwToolExit(null, exitCode: 123);
       };
       appStopper = expectAsync1((DriveCommand command) async {
@@ -373,7 +361,7 @@ void main() {
         testApp = fs.path.join(tempDir.path, 'test', 'e2e.dart');
         testFile = fs.path.join(tempDir.path, 'test_driver', 'e2e_test.dart');
 
-        testRunner = (List<String> testArgs, Map<String, String> environment) async {
+        testRunner = (List<String> testArgs, String observatoryUri) async {
           throwToolExit(null, exitCode: 123);
         };
         appStopper = expectAsync1(
@@ -506,7 +494,7 @@ void main() {
         testApp = fs.path.join(tempDir.path, 'test', 'e2e.dart');
         testFile = fs.path.join(tempDir.path, 'test_driver', 'e2e_test.dart');
 
-        testRunner = (List<String> testArgs, Map<String, String> environment) async {
+        testRunner = (List<String> testArgs, String observatoryUri) async {
           throwToolExit(null, exitCode: 123);
         };
         appStopper = expectAsync1(
